Implemented the encryption and decryption
Made some modifications so the app doesn't stop when choosing an invalid option,  implemented encryption , checks for the rooms names and I think the problem with multiple clients is because the same machine is trying to connect (still need to test that out).
We still need to implement a way to check the passwords of the rooms and a way for the client to input the password of the room he tries to connect to.